refactor: rename contact module to support and update package names, artifactId, and references
