+++ support/src/main/java/pt/estga/support/config/ContactAutoConfiguration.java
@@ -1,4 +1,4 @@
-package pt.estga.contact.config;
+package pt.estga.support.config;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.persistence.autoconfigure.EntityScan;
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @AutoConfiguration
-@ComponentScan("pt.estga.contact")
-@EnableJpaRepositories("pt.estga.contact.repositories")
+@ComponentScan("pt.estga.support")
+@EnableJpaRepositories("pt.estga.support.repositories")
 @EntityScan("pt.estga.contact.entities")
 public class ContactAutoConfiguration {
 }
