+++ support/src/main/java/pt/estga/support/controllers/ContactRequestAdminController.java
@@ -1,4 +1,4 @@
-package pt.estga.contact.controllers;
+package pt.estga.support.controllers;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -6,10 +6,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import pt.estga.contact.enums.ContactStatus;
-import pt.estga.contact.entities.ContactRequest;
-import pt.estga.contact.services.ContactRequestQueryService;
-import pt.estga.contact.services.ContactRequestService;
+import pt.estga.support.enums.ContactStatus;
+import pt.estga.support.entities.ContactRequest;
+import pt.estga.support.services.ContactRequestQueryService;
+import pt.estga.support.services.ContactRequestService;
 import pt.estga.sharedweb.models.PagedRequest;
 
 @RestController
@@ -17,7 +17,7 @@ import pt.estga.sharedweb.models.PagedRequest;
 @RequiredArgsConstructor
 @Tag(name = "Admin Contact Requests", description = "Admin endpoints for contact requests.")
 @PreAuthorize("hasRole('MODERATOR')")
-public class AdminContactRequestController {
+public class ContactRequestAdminController {
 
     private final ContactRequestQueryService queryService;
     private final ContactRequestService service;
