+++ support/src/main/java/pt/estga/support/controllers/ContactRequestPublicController.java
@@ -1,4 +1,4 @@
-package pt.estga.contact.controllers;
+package pt.estga.support.controllers;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -13,16 +13,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pt.estga.shared.interfaces.AuthenticatedPrincipal;
-import pt.estga.contact.dtos.ContactRequestDto;
-import pt.estga.contact.entities.ContactRequest;
-import pt.estga.contact.services.ContactRequestService;
-import pt.estga.contact.services.ContactRequestQueryService;
+import pt.estga.support.dtos.ContactRequestDto;
+import pt.estga.support.entities.ContactRequest;
+import pt.estga.support.services.ContactRequestService;
+import pt.estga.support.services.ContactRequestQueryService;
 
 @RestController
 @RequestMapping("/api/v1/public/contact-requests")
 @RequiredArgsConstructor
 @Tag(name = "Public Contact Requests", description = "Public endpoints for contact requests.")
-public class PublicContactRequestController {
+public class ContactRequestPublicController {
 
     private final ContactRequestService service;
     private final ContactRequestQueryService queryService;
