+++ support/src/main/java/pt/estga/support/dtos/ContactRequestDto.java
@@ -1,4 +1,4 @@
-package pt.estga.contact.dtos;
+package pt.estga.support.dtos;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
