+++ support/src/main/java/pt/estga/support/entities/ContactRequest.java
@@ -1,9 +1,9 @@
-package pt.estga.contact.entities;
+package pt.estga.support.entities;
 
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
-import pt.estga.contact.enums.ContactStatus;
+import pt.estga.support.enums.ContactStatus;
 import pt.estga.user.entities.User;
 
 import java.time.Instant;
