+++ support/src/main/java/pt/estga/support/enums/ContactStatus.java
@@ -1,4 +1,4 @@
-package pt.estga.contact.enums;
+package pt.estga.support.enums;
 
 public enum ContactStatus {
     PENDING,
